added rules for linting and prettier
