+++ README.md
@@ -1,0 +1,23 @@
+**To run locally**
+
+1. Install dependencies:
+
+```s
+npm install
+```
+
+2. Start environment
+
+```s
+npm start
+```
+
+3. Before merging to master, check errors and prettier rules
+
+```s
+npm lint
+```
+
+**To test with device in development mode or production mode**
+
+1. You need to connect to MongoAtlas by providing the MONGOURI
