+++ server/.eslintrc
@@ -1,0 +1,10 @@
+{
+  "extends": "airbnb-base",
+  "env": {
+    "node": true,
+    "es6": true
+  },
+  "plugins": [
+    "import"
+  ]
+}
