+++ server/.prettierrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+    bracketSpacing: true,
+    singleQuote: true,
+    trailingComma: 'all',
+};
