--- server/config/key.js
@@ -1,5 +1,0 @@
-if (process.env.NODE_ENV === "production") {
-    module.exports = require("./production");
-} else {
-    module.exports = require("./development");
-}
