--- server/models/User.js
@@ -1,9 +1,0 @@
-const mongoose = require ('mongoose');
-
-const userSchema = new mongoose.Schema({
-    username: {type: String, required: true, trim:true},
-    email: {type: String, required: true, trim:true},
-    password: {type: String, required: true}
-})
-
-module.exports = mongoose.model('User', userSchema);
