+++ server/src/config/key.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./production');
+} else {
+  module.exports = require('./development');
+}
