+++ server/src/routes/verifyToken.js
@@ -1,0 +1,15 @@
+const jwt = require('jsonwebtoken');
+
+module.exports = function (req, res, next) {
+  const token = req.header('auth-token');
+  if (!token) return res.status(401).send('Access denied');
+
+  // verifying token
+  try {
+    const verified = jwt.verify(token, 'SECRET_KEY');
+    req.user = verified;
+    next();
+  } catch (error) {
+    res.status(400).send('Invalid token');
+  }
+};
